Cleaned up Cleaned up code
